STAF-220: make sure POST / PATCH assignments work
- The route schema needed an update to use the allow the resource
relation object instead of the id (BelongsToOne). E.g: instead of
allowing employee_id. The schema should allow for an employee object
with an id property since this is how the relationships are
deserialized

- The objection model relationship for BelongsToOne *MUST* be singular,
an Assignment has a BelongsToOne relation with a single *employee* and a
single *role*. Updating the model allows objection to handle the updated
payload properly.